rename StoresSpec to CheckpointsSpec factor out Stores generators code
